refactor: 인스턴스 상세 컴포넌트 구조 개선
