+++ src/api/index.ts
@@ -8,5 +8,5 @@ const create = (baseURL: string, options?: CreateAxiosDefaults): AxiosInstance =
   return instance
 }
 
-// json-server 엔드포인트를 생성
+// json-server 엔드포인트 생성
 export const instancesAxios = create('http://localhost:3000')
