+++ src/components/instance/detail/components/index.ts
@@ -1,0 +1,3 @@
+export { default as BasicInfo } from './BasicInfo.vue'
+export { default as NetworkInfo } from './NetworkInfo.vue'
+export { default as ResourceInfo } from './ResourceInfo.vue'
