+++ src/components/instance/detail/modals/EditBasicInfoModal.vue
