+++ src/components/instance/detail/modals/EditResourceInfoModal.vue
