+++ src/components/instance/detail/modals/index.ts
@@ -1,0 +1,3 @@
+export { default as EditBasicInfoModal } from './EditBasicInfoModal.vue'
+export { default as EditNetworkInfoModal } from './EditNetworkInfoModal.vue'
+export { default as EditResourceInfoModal } from './EditResourceInfoModal.vue'
